Upon my first review, I have found that reflection and list access are the immediate worst offenders at this point.  I ran a read & write operation as indicated below and found the following with a CAD DXF drawing that has roughly 22,000 items (Blocks & simple entities).  I would attach the DXF, but it is a private project file.

For this first pass, I tired to leave the base code alone as much as possible as my initial work on refactoring the Map classes to use static instances became wide reaching very fast.

With the current changes, we get a performance increase of roughly 5x on the first read and then 6.5x faster on subsequent reads over the baseline. The write performance is also now 4.7x faster.

Ran into a small reading issue with `DxfSectionReaderBase.readMapped` method where it was trying to read a `DxfReferenceType.Handle` for a FontStyle but the file actually contains a string value.  I'll address that in a separate PR.

#### Changes
- Reflection used for setting properties has now been changed out for using Expressions for generating cached setters & getters for properties via the `PropertyExpression` class.
- CadObjectCollection<T>.Entities is now a `HashMap<T>` due to `Contains()` lookups during addition of entities.
- `DxfMap` and `DxfClassMap` now both cache their results of the `Create<T>()` method.  Caches are cleared via the `ClearCache()` method on both types.  Caching these items reduces the reflection on the first read/write significantly and all future reads and writes have all the maps already cached and are thus faster.

``` csharp
var sw = Stopwatch.StartNew();
Console.WriteLine("Startup Read & Write");

var document = DxfReader.Read("X-Base.dxf", (o, e) => { });
Console.WriteLine($"[{sw.ElapsedMilliseconds:0,000}ms] Completed Read");
sw.Restart();
DxfWriter.Write("X-Base-output.dxf", document, false);
Console.WriteLine($"[{sw.ElapsedMilliseconds:0,000}ms] Completed Write");

Console.WriteLine($"Completed startup Read & Write.");
Console.WriteLine();

sw.Restart();
for (int i = 0; i < 1; i++)
{
    sw.Restart();
    Console.WriteLine($"Started Subsequent Read & Write number {i + 1}");
    document = DxfReader.Read("X-Base.dxf", (o, e) => { });
    Console.WriteLine($"[{sw.ElapsedMilliseconds:0,000}ms] Completed Read");

    sw.Restart();
    DxfWriter.Write("X-Base-output.dxf", document, false);
    Console.WriteLine($"[{sw.ElapsedMilliseconds:0,000}ms] Completed Write");

    Console.WriteLine($"Completed Subsequent Read & Write number {i + 1}");
    Console.WriteLine();
}
```

## Baseline Performance
``` 
Startup Read & Write
[5,693ms] Completed Read
[2,264ms] Completed Write
Completed startup Read & Write.

Started Subsequent Read & Write number 1
[5,534ms] Completed Read
[2,278ms] Completed Write
Completed Subsequent Read & Write number 1
```
#### Dot Trace Performance
![image](https://user-images.githubusercontent.com/224169/181372623-587f2123-c691-43bc-ac91-0acb548eec0f.png)


## After PR Modifications
```
Startup Read & Write
[1,114ms] Completed Read
[0,486ms] Completed Write
Completed startup Read & Write.

Started Subsequent Read & Write number 1
[0,872ms] Completed Read
[0,495ms] Completed Write
Completed Subsequent Read & Write number 1
```
#### Dot Trace Performance
![image](https://user-images.githubusercontent.com/224169/181374418-83af4e20-a4b1-4f48-9eb8-4ddb68c1d3f1.png)